Revert PR #91, move Module.php out of src/
This was mainly to appease composer's PSR0 autoloader, however we should
not have to alter module structure (especially the one and only single
requirement for a ZF2 module) simply for the sake of composer. Besides,
composer supports classmaps which resolves the issue just fine.
